Add styles and fix favorite option

+++ src/app/recipe-detail/recipe-detail.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { ActivatedRoute } from '@angular/router';
+import {ActivatedRoute, Router} from '@angular/router';
 import { RecipeService } from '../recipe.service';
 import { DomSanitizer, SafeHtml } from '@angular/platform-browser';
 import {AuthService} from "../auth.service";
@@ -17,9 +17,11 @@ export class RecipeDetailComponent implements OnInit {
   user$: Observable<firebase.User | null>;
   isFavorite: boolean = false;
   isProfilePage: boolean = false;
+  searchActive: boolean = false;
 
   constructor(
     private route: ActivatedRoute,
+    private router: Router,
     private recipeService: RecipeService,
     private sanitizer: DomSanitizer,
     private authService: AuthService
@@ -43,6 +45,11 @@ export class RecipeDetailComponent implements OnInit {
     this.authService.logout();
   }
 
+  showCarousel(): void {
+    this.isProfilePage = false;
+    this.router.navigate(['/recipes']);
+  }
+
   goToProfile(): void {
     this.isProfilePage = true;
   }
@@ -50,10 +57,12 @@ export class RecipeDetailComponent implements OnInit {
   loadRecipeDetails(recipeId: string): void {
     this.recipeService.getRecipeById(recipeId).subscribe(
       (response) => {
-        console.log('Recipe details fetched:', response); // Log fetched recipe
         this.recipe = response;
+        if (!this.recipe.image) {
+          this.recipe.image = 'assets/cooker.png';
+        }
         this.sanitizedInstructions = this.sanitizer.bypassSecurityTrustHtml(this.recipe.instructions);
-        this.checkIfFavorite(); // Check if the recipe is favorite after loading it
+        this.checkIfFavorite();
       },
       (error) => {
         console.error('Error fetching recipe details', error);
@@ -65,11 +74,19 @@ export class RecipeDetailComponent implements OnInit {
     this.authService.user$.subscribe((user: firebase.User | null) => {
       if (user) {
         if (this.isFavorite) {
-          this.authService.removeFavoriteRecipe(user.uid, this.recipe)
-            .then(() => this.isFavorite = false);
+          if (this.recipe && this.recipe.id) {
+            this.authService.removeFavoriteRecipe(user.uid, this.recipe)
+              .then(() => this.isFavorite = false);
+          } else {
+            console.error('Recipe is undefined or does not have an id');
+          }
         } else {
-          this.authService.addFavoriteRecipe(user.uid, this.recipe)
-            .then(() => this.isFavorite = true);
+          if (this.recipe && this.recipe.id) {
+            this.authService.addFavoriteRecipe(user.uid, this.recipe)
+              .then(() => this.isFavorite = true);
+          } else {
+            console.error('Recipe is undefined or does not have an id');
+          }
         }
       }
     });
@@ -79,11 +96,11 @@ export class RecipeDetailComponent implements OnInit {
     this.authService.getFavorites().subscribe((favorites) => {
       if (this.recipe && this.recipe.id) {
         this.isFavorite = favorites.some((r: any) => r.id === this.recipe.id);
-        console.log('Favorite status:', this.isFavorite);
       } else {
         console.error('Recipe is undefined or does not have an id');
       }
     });
   }
+
 
 }
